adding interpolation and notebook
